# api/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Hub API
  version: 0.3.0
  summary: One chat interface for many AI agents. Browser ↔ API contract, plus the API → agent payload.
  description: |
    The contract between the Agent Hub web app and its FastAPI service (build plan step 5).
    Decisions behind it: `docs/ARCHITECTURE.md` (D1–D23). Send/stream rules and timers: `docs/SEND_MESSAGE.md`.
    Tables: `docs/DATA_MODEL.md`.

    ## Conventions (D23)
    - All routes are under `/v1`. JSON is camelCase. IDs are UUID strings. Timestamps are RFC 3339 in UTC.
    - Agents are addressed by `slug`; slugs never change.
    - Every response carries `X-Request-Id`. Send it back in bug reports; errors include it as `requestId`.
    - Operations are marked `x-scope: v1` (P0, needed for launch) or `x-scope: P1` (designed, built after v1).

    ## Auth (D8)
    1. The browser gets a Google ID token and calls `POST /v1/auth/google` once.
    2. The API returns a short-lived access token (~15 min, keep it in memory, send it as `Authorization: Bearer`)
       and sets a rotating refresh cookie `ah_refresh` (`HttpOnly; Secure; SameSite=Strict; Path=/v1/auth`).
    3. On app load, and whenever a call returns `401 token_expired`, call `POST /v1/auth/refresh`, then repeat the call
       (resends are safe: see Idempotency). `401 session_expired` means signing in again.
    `/v1/auth/*` also checks the `Origin` header. Every other route needs the bearer token, except `/v1/health` and
    `/v1/public/*`.

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /v1/agents:
    get:
      tags: [Agents]
      operationId: listAgents
      summary: The catalog
      description: Listed, not retired agents, in catalog order (the descriptor's `sortOrder`, then name). Not paginated; the registry is small. Hidden agents (the Scenario Agent) are included everywhere except production.
      x-scope: v1
      responses:
        '200':
          description: Agents.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agent'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/agents/{slug}:
    get:
      tags: [Agents]
      operationId: getAgent
      summary: One agent, including retired and hidden ones
      description: Retired agents are returned (with `retiredAt`) so their old sessions stay readable. An unknown slug is `404 agent_not_found`; one that doesn't match the slug pattern is `422 invalid_request`.
      x-scope: v1
      parameters:
        - $ref: '#/components/parameters/AgentSlug'
      responses:
        '200':
          description: The agent.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Invalid'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from `/v1/auth/google` or `/v1/auth/refresh`, valid about 15 minutes.
    refreshCookie:
      type: apiKey
      in: cookie
      name: ah_refresh
      description: Rotating refresh token, `HttpOnly; Secure; SameSite=Strict; Path=/v1/auth`, 30 days.
  headers:
    XRequestId:
      description: Unique id of this request; also `requestId` in errors.
      schema:
        type: string
...
  parameters:
    AgentSlug:
      name: slug
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-z0-9]+(-[a-z0-9]+)*$'
      example: research-analyst
...
  responses:
...
    Unauthorized:
      description: '`token_invalid`, `token_expired` (refresh, then repeat), `session_expired` (sign in again) or `invalid_google_token`.'
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://fleet.qucoon.com/errors/token_expired
            title: Access token expired
            status: 401
            code: token_expired
            requestId: req_7a1f0b
            retryable: true
...
    NotFound:
      description: The resource doesn't exist, was deleted, or belongs to someone else (`agent_not_found`, `session_not_found`, `message_not_found`, `approval_not_found`, `file_not_found`, `artifact_not_found`, `share_not_found`).
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://fleet.qucoon.com/errors/session_not_found
            title: Session not found
            status: 404
            code: session_not_found
            requestId: req_19be02
            retryable: false
...
    Invalid:
      description: 'Validation failed: `invalid_request` (with `errors`), `empty_message`, `message_too_long`, `too_many_files`, `file_too_large`, `file_type_not_allowed`, `attachments_not_supported`, `upload_mismatch`, `confirmation_required`.'
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://fleet.qucoon.com/errors/file_too_large
            title: File is too large
            status: 422
            detail: ct-scan-full.tiff is 86 MB; the limit is 20 MB.
            code: file_too_large
            requestId: req_5d0e71
            retryable: false
...
  schemas:
...
    Uuid:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
    NullableTimestamp:
      type: [string, 'null']
      format: date-time

    # ------------------------------------------------------------ Errors
    ErrorCode:
      type: string
      description: |
        Stable machine-readable codes. HTTP status in brackets; "event" means it appears in `run.failed` or `Message.error`.
        - `invalid_request` [422] body or query failed validation (`errors` lists the fields)
        - `empty_message` [422] no text, files or answer
        - `message_too_long` [422] text over 32,000 characters
        - `too_many_files` [422] more than 10 attachments, or more than the agent allows
        - `file_too_large` [422], `file_type_not_allowed` [422], `attachments_not_supported` [422] the agent takes no files
        - `upload_mismatch` [422] the object in S3 doesn't match the declared size or type
        - `confirmation_required` [422] "Delete all" without the typed confirmation
        - `token_invalid` [401] missing or bad access token
        - `token_expired` [401] refresh, then repeat the request
        - `invalid_google_token` [401] the Google ID token failed verification
        - `session_expired` [401] refresh token expired, revoked or reused; sign in again
        - `origin_not_allowed` [403], `not_invited` [403], `account_disabled` [403]
        - `forbidden` [403] signed in, but the action needs a role the user doesn't have
        - `agent_not_found`, `session_not_found`, `message_not_found`, `approval_not_found`, `file_not_found`,
          `artifact_not_found`, `share_not_found` [404]
        - `not_found` [404] no such route; `method_not_allowed` [405] the route exists but not for this method (`Allow` lists the methods)
        - `conflict` [409] generic; the specific codes below are used where one exists
        - `run_in_progress` [409] a reply is streaming in this session
        - `agent_retired` [409] the agent was removed; its sessions are read-only
        - `not_retryable` [409] retry on a message that isn't stopped, failed or interrupted
        - `approval_not_pending` [409] the approval was already decided, expired or cancelled
        - `question_not_open` [409] the answered question was already answered or skipped
        - `idempotency_conflict` [409] a `clientMessageId` reused with different content
        - `file_not_ready` [409] sending a file that isn't `uploaded`, or completing one before the bytes arrived
        - `file_attached` [409] deleting a file that was already sent
        - `payload_too_large` [413] the request body is over the limit (files go to S3 directly, D7)
        - `too_many_runs` [429] three replies already running for this user
        - `agent_unavailable` [503, event] agent offline, or AgentCore unreachable
        - `identity_provider_unavailable` [503] Google's signing keys couldn't be fetched during sign-in; retry
        - `shutting_down` [503] the API task is draining for a deploy; retry (`retryAfter`) (P2)
        - `agent_error` [event] the agent failed mid-run
        - `rate_limited` [event] AgentCore or the model throttled the call (`retryAfter`)
        - `run_time_limit` [event] the 15-minute run cap was hit; status `interrupted`
        - `run_interrupted` [event] the task running the reply died (deploy or crash); status `interrupted`
        - `internal_error` [500, event]
      enum:
        - invalid_request
        - empty_message
        - message_too_long
        - too_many_files
        - file_too_large
        - file_type_not_allowed
        - attachments_not_supported
        - upload_mismatch
        - confirmation_required
        - token_invalid
        - token_expired
        - invalid_google_token
        - session_expired
        - origin_not_allowed
        - not_invited
        - account_disabled
        - forbidden
        - agent_not_found
        - session_not_found
        - message_not_found
        - approval_not_found
        - file_not_found
        - artifact_not_found
        - share_not_found
        - not_found
        - method_not_allowed
        - conflict
        - run_in_progress
        - agent_retired
        - not_retryable
        - approval_not_pending
        - question_not_open
        - idempotency_conflict
        - file_not_ready
        - file_attached
        - payload_too_large
        - too_many_runs
        - agent_unavailable
        - identity_provider_unavailable
        - shutting_down
        - agent_error
        - rate_limited
        - run_time_limit
        - run_interrupted
        - internal_error
    Problem:
      type: object
      description: RFC 9457 problem details with Agent Hub extensions.
      required: [type, title, status, code, requestId, retryable]
      properties:
        type:
          type: string
          format: uri
          description: '`https://fleet.qucoon.com/errors/<code>`'
        title:
          type: string
          description: Short, human-readable, safe to show.
        status:
          type: integer
          description: HTTP status. In stream events and stored errors, the status the error would have had (500 for agent errors).
        detail:
          type: string
        code:
          $ref: '#/components/schemas/ErrorCode'
        requestId:
          type: string
          description: Shown in the UI as "req_8f2a" for bug reports.
        retryable:
          type: boolean
          description: Whether repeating the same action can succeed.
        retryAfter:
          type: integer
          minimum: 0
          description: Seconds to wait before retrying, when known.
        errors:
          type: array
          description: Field errors, for `invalid_request`.
          items:
            type: object
            required: [path, message]
            properties:
              path:
                type: string
                description: JSON pointer into the request, e.g. `/fileIds/3`.
              message:
                type: string
        email:
          type: string
          format: email
          description: For `not_invited`.
        messageId:
          $ref: '#/components/schemas/Uuid'
        approval:
          $ref: '#/components/schemas/Approval'
        limits:
          $ref: '#/components/schemas/AttachmentLimits'

    # ------------------------------------------------------------ Auth and users
...
    AgentColor:
      type: string
      enum: [red, blue, green, purple, amber, grey]
      description: Maps to the `$agent-<color>-bg/fg` tokens. Identity only.
    AgentStatus:
      type: string
      enum: [online, degraded, offline]
      description: Health. Set by hand in v1 (D23); computed in P1 (F23).
    AgentStage:
      type: string
      enum: [beta, stable]
      description: Maturity, shown as a separate Beta tag.
    AgentRef:
      type: object
      description: The agent fields shown next to sessions, artifacts and search results.
      required: [id, slug, name, icon, color, status, stage, retired]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        slug:
          type: string
        name:
          type: string
        icon:
          type: string
          description: lucide icon name.
        color:
          $ref: '#/components/schemas/AgentColor'
        status:
          $ref: '#/components/schemas/AgentStatus'
        stage:
          $ref: '#/components/schemas/AgentStage'
        retired:
          type: boolean
    Agent:
      type: object
      required:
        - id
        - slug
        - name
        - description
        - tagline
        - greeting
        - icon
        - color
        - stage
        - status
        - statusChangedAt
        - framework
        - runtimeType
        - version
        - deployedAt
        - details
        - disclaimer
        - capabilities
        - tools
        - starters
        - retiredAt
        - myStats
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        slug:
          type: string
          example: coding-agent
        name:
          type: string
          example: Coding Agent
        description:
          type: string
        tagline:
          type: [string, 'null']
          description: Short line for the mobile catalog. Falls back to `description`.
        greeting:
          type: [string, 'null']
          description: New-session heading, e.g. "What are we building today?".
        icon:
          type: string
          description: lucide icon name.
        color:
          $ref: '#/components/schemas/AgentColor'
        stage:
          $ref: '#/components/schemas/AgentStage'
        status:
          $ref: '#/components/schemas/AgentStatus'
        statusChangedAt:
          $ref: '#/components/schemas/NullableTimestamp'
        framework:
          type: [string, 'null']
          description: Display only, e.g. `strands`, `langgraph`.
        runtimeType:
          type: string
          enum: [agentcore, http]
        version:
          type: [string, 'null']
          description: Deployed version; a change between two replies shows the "Agent updated" divider.
        deployedAt:
          $ref: '#/components/schemas/NullableTimestamp'
        details:
          $ref: '#/components/schemas/AgentDetails'
        disclaimer:
          type: [string, 'null']
          description: Shown under the composer after "{name} can make mistakes."
        capabilities:
          $ref: '#/components/schemas/AgentCapabilities'
        tools:
          type: array
          items:
            $ref: '#/components/schemas/AgentTool'
        starters:
          type: array
          maxItems: 6
          items:
            $ref: '#/components/schemas/Starter'
        retiredAt:
          $ref: '#/components/schemas/NullableTimestamp'
        myStats:
          type: object
          description: For the signed-in user.
          required: [sessionCount, lastActiveAt]
          properties:
            sessionCount:
              type: integer
              minimum: 0
            lastActiveAt:
              $ref: '#/components/schemas/NullableTimestamp'
    AgentDetails:
      type: object
      description: Display-only facts for the agent detail page.
      properties:
        model:
          type: string
          example: Claude Sonnet 5
        region:
          type: string
          example: us-east-1
        frameworkVersion:
          type: string
          example: Strands 1.8
        runtimeLabel:
          type: string
          description: Shortened runtime ARN or endpoint for display.
          example: arn:…:runtime/coding-7Hq2
    AgentCapabilities:
      type: object
      description: What the UI shows for this agent. The UI hides what isn't supported (no "+" without attachments).
      required: [attachments, artifacts, approvals, questions]
      properties:
        attachments:
          $ref: '#/components/schemas/AttachmentLimits'
        artifacts:
          type: boolean
        approvals:
          type: boolean
        questions:
          type: boolean
    AttachmentLimits:
      type: object
      required: [enabled, contentTypes, maxFileBytes, maxFiles]
      properties:
        enabled:
          type: boolean
        contentTypes:
          type: array
          items:
            $ref: '#/components/schemas/UploadContentType'
        maxFileBytes:
          type: integer
          maximum: 20971520
        maxFiles:
          type: integer
          maximum: 10
    UploadContentType:
      type: string
      enum: [application/pdf, image/png, image/jpeg, image/webp, image/gif, text/csv, text/plain]
    AgentTool:
      type: object
      required: [name, description, requiresApproval]
      properties:
        name:
          type: string
          example: run_command
        description:
          type: string
        requiresApproval:
          type: boolean
          description: Shown as "Asks first"; the API enforces it (D19).
    Starter:
      type: object
      required: [title, prompt]
      properties:
        title:
          type: string
          example: Find why a test is flaky
        description:
          type: string
          example: Runs it 20× and reads the failures
        prompt:
          type: string
          description: Text placed in the composer when the starter is picked.

    # ------------------------------------------------------------ Sessions
...
    ApprovalStatus:
      type: string
      enum: [pending, approved, denied, expired, cancelled]
    Approval:
      type: object
      required: [id, sessionId, messageId, toolCallId, toolName, status, display, arguments, requestedAt, expiresAt, decidedAt, decisionReason, decidedOn, executedAt]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        sessionId:
          $ref: '#/components/schemas/Uuid'
        messageId:
          $ref: '#/components/schemas/Uuid'
        toolCallId:
          $ref: '#/components/schemas/Uuid'
        agent:
          $ref: '#/components/schemas/AgentRef'
        toolName:
          type: string
          example: execute_transfer
        status:
          $ref: '#/components/schemas/ApprovalStatus'
        display:
          $ref: '#/components/schemas/ApprovalDisplay'
        arguments:
          type: object
          description: The tool arguments the API validated. The card shows `display`, which is derived from these, never model prose.
        requestedAt:
          $ref: '#/components/schemas/Timestamp'
        expiresAt:
          $ref: '#/components/schemas/Timestamp'
        decidedAt:
          $ref: '#/components/schemas/NullableTimestamp'
        decisionReason:
          type: [string, 'null']
        decidedOn:
          type: [string, 'null']
          description: Short device label from the deciding request ("your phone") for "Approved on your phone".
        executedAt:
          $ref: '#/components/schemas/NullableTimestamp'
    ApprovalDisplay:
      type: object
      required: [title, fields]
      properties:
        title:
          type: string
          example: Transfer NGN 50,000 to Ada O.
        fields:
          type: array
          items:
            type: object
            required: [label, value]
            properties:
              label:
                type: string
                example: Amount
              value:
                type: string
                example: NGN 50,000.00
        note:
          type: string
          example: Transfers can't be reversed once sent. Nothing moves until you approve.
        approveLabel:
          type: string
          example: Approve transfer
...